preparation for 5'/3'-prime modifications
Some modifications are located on the 5' and 3'-end of any nucleotide. Among these are modified nucleotides as well, which are not linked via 5'/3'-phosphate so they are not part of the RNA chain. This represents a challenge for encoding, since they can only occur on the first/last position. This require special encoding rules.

For now, these types of modifications are included with an empty letter, which removes them from the alphabet. However, they are available for sanitization, so that a lossy conversion from character string is possible, without additional function calls by the user.